refactor(protocol): move catalogCommands into the package
slashCommandInvocation and visibleCatalogCommands are shared by the web
palette, the composer's inline slash menu and the native command catalog,
so they belong in the package rather than under shell/palette. No logic
changes: the module moves, its types.gen import becomes intra-package,
and every importer's specifier is repointed.
