+++ cmd/evener-hub/frontend/src/protocol/README.md
@@ -10,7 +10,8 @@ attachment markers into prose at send, the thread view model and its
 notification reducer, the activity tree parser, merge and disclosure rules,
 the job log tail parser, the send/queue availability table, the
 send/steer/queue/drain routing decisions a composer makes off it, the stable
-delegate status rule, the display formatters both apps render counts,
+delegate status rule, the slash invocation and catalog visibility rules the
+palette and composer share, the display formatters both apps render counts,
 durations and clock times with, and the doc-pane URL builders, which hang
 their hrefs off a base origin the host supplies (empty for a same-origin web
 page). The doc-pane data layer is published at the `./docContent` subpath as
