+++ cmd/evener-hub/frontend/src/protocol/scripts/qualify-package.mjs
@@ -54,6 +54,7 @@ async function qualify() {
     "docContent",
     "submitRouting",
     "displayFormat",
+    "catalogCommands",
   ];
   // Every runtime export of the package root. The root's generated consumer
   // programs are built from this one list, so an export the entry point stops
@@ -137,6 +138,8 @@ async function qualify() {
     "firstLine",
     "splitMandate",
     "plainQuoteLine",
+    "slashCommandInvocation",
+    "visibleCatalogCommands",
   ];
   // One exported type per shipped module that declares any, so the declaration
   // check covers each module's packed .d.ts and not just its runtime half.
@@ -220,6 +223,8 @@ assert.equal(client.formatElapsed(65000), "1m05s");
 assert.equal(client.firstLine("\\n  hello  \\n", 20), "hello");
 assert.deepEqual(client.splitMandate("first\\n\\nrest"), { first: "first", rest: "rest" });
 assert.equal(client.plainQuoteLine("# Title\\n**bold** line"), "bold line");
+assert.equal(client.slashCommandInvocation({ name: "plan", source: "plugin", pluginName: "acme" }), "/acme:plan");
+assert.deepEqual(client.visibleCatalogCommands([{ name: "plan", source: "plugin", pluginName: "acme" }], new Set()), []);
 const activity = new client.ActivityList({ request: async () => ({}), onNotification: () => () => {} }, "ref", "thread");
 assert.equal(activity.getSnapshot().tree, null);
 `;
