+++ cmd/evener-hub/frontend/src/protocol/tsconfig.build.json
@@ -33,6 +33,7 @@
     "stableDelegate.ts",
     "docContent.ts",
     "submitRouting.ts",
-    "displayFormat.ts"
+    "displayFormat.ts",
+    "catalogCommands.ts"
   ]
 }
